requirements.txt
opencv-python
opencv-contrib-python
pillow
numpy
scipy
mtcnn
keras_vggface
